Minor fixes Cleanup
